supersieve-backend/requirements.txt
django~=5.1
django-restframework
django_summernote
djangorestframework-simplejwt
django_summernote
whitenoise
django-utils-six
django-utils
scikit-learn
Pillow
psycopg2-binary
pymysql
sqlparse
pytz
asgiref
certifi
social-auth-app-django
pdfplumber
regex
python-docx
python-dotenv
django-cors-headers
django-filter
html2text
beautifulsoup4
outlook_msg
drf-url-filters
django-environ
mammoth
twilio~=9.2.4
plivo~=4.55.1
Faker
djangorestframework~=3.15.2
six~=1.16.0